Build(Navigation): Install Dependencies
Install required Dependencies for Navigation, react-navigation/elements (set of components), react-navigation/native (core utilities), react-navigation/native-stack (building blocks and shared foundations for navigators). These dependencies all involved in he navigation.

I am taking this navigation a step further by using Drag & Drop which is a device interaction essential fro the project. Required dependencies being react-native-reanimated-dnd (This is for animations of the dragging) and react-native-gesture-handler (This is for the actual gestures of a user dragging each component). These two dependencies are essential for the drag and drop function