Feat(Backend): Setup Backend Data
Preparing my data in order to send it to Firebase, how I want the data to be structured within Firestore then,
explicitly describing the type want the data to be,
I also configured how the data should be send example have added logic to calculate duration of trip and convert it into seconds,
Once trip details have started then added, the status of trip is set to Ongoing initially, then added an update logic for changing status from Ongoing to completed or cancel status. the completed status is applied only once user has completed the puzzle in the puzzle screen, startTripId is exported so it can be parsed into other screens.