Feat(UseCases): Add renderEmpty
the renderEmpty is to place a message for user to add friends, this will mostly display is for newly registered users, although if an existing user removes their friends this should also display.

Using FlatList as this is most applicable due to fact that this accommodates users with many friends as it helps with memory loading so app dos not start to slow down due to a lot of rendering thus most applicable to use FlatList.

Added inlineStyling to make flatlist and renderEmpty message look clean